SCRUM-97 refactor: Refactor login logic into a reusable custom hook.
